% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ace.test.R
\name{ace.test}
\alias{ace.test}
\title{ACE permutation test of independence}
\usage{
ace.test(x, y = NULL, nperm = 999)
}
\arguments{
\item{x}{a numeric vector, or a matrix or data frame with two columns.}

\item{y}{a vector with same length as x. Default is NULL.}

\item{nperm}{number of permutations. Default is 999.}
}
\value{
a list containing the following:
\itemize{
  \item{\code{ace}}{ The value of the test statistic.}
  \item{\code{pval}}{ The *p*-value of the test.}
}
}
\description{
Performs a permutation test of independence or association. The
  alternative hypothesis is that there is a x and y are dependent. 
  
Code authored by Hajo Holzmann, Bernhard Klar.
}
\examples{

n <- 200
x <- matrix(rnorm(n*2), n)
nu <- 2
y <- x / sqrt(rchisq(n, nu)/nu) #multivariate t
cor.test(y[,1], y[,2], method = "spearman")
ace.test(y)
}
\references{
Holzmann, H., Klar, B. 2025. "Lancaster correlation - a new dependence measure
linked to maximum correlation". Scandinavian Journal of Statistics.
52(1):145-169 <doi:10.1111/sjos.12733>
}
