% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ace.test.R
\name{plot.acetest}
\alias{plot.acetest}
\title{ACE permutation histogram}
\usage{
\method{plot}{acetest}(
  x,
  acol = "blue",
  xlim = c(min(x$tp), max(c(x$tp, ceiling(x$ace * 10)/10))),
  col = "black",
  breaks = 100,
  main = "ACE Correlation Permutations",
  xlab = bquote(rho(.(x$xname), .(x$yname))),
  lwd = 2,
  ...
)
}
\arguments{
\item{x}{the ace.test object to print}

\item{acol}{color of the point estimate of correlation}

\item{xlim}{xlimit of histogram}

\item{col}{color of histogram bars}

\item{breaks}{number of breaks. default to 100}

\item{main}{main title of plot}

\item{xlab}{x-axis label}

\item{lwd}{line width of point estimate}

\item{...}{additional arguments to send to hist}
}
\description{
An S3 function to produce a summary 
  plot of the results of an ace.test.
}
\examples{

n <- 200
x <- matrix(rnorm(n*2), n)
nu <- 2
y <- x / sqrt(rchisq(n, nu)/nu) #multivariate t
plot(ace.test(y))

}
