% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ace.test.R
\name{print.acetest}
\alias{print.acetest}
\title{ACE permutation test summary}
\usage{
\method{print}{acetest}(x, ...)
}
\arguments{
\item{x}{the ace.test object to print}

\item{...}{additional arguments to send to cat}
}
\value{
original object
}
\description{
An S3 function to produce a summary 
  of the results of an ace.test.
}
