% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ace.test.R
\name{summary.ace.test}
\alias{summary.ace.test}
\alias{print.ace.test}
\title{ACE permutation test summary}
\usage{
\method{summary}{ace.test}(object, ..., digits)

\method{print}{ace.test}(x, ...)
}
\arguments{
\item{object}{the test to summarize}

\item{...}{additional arguments to send to cat}

\item{digits}{Number of significant digits to round too.}

\item{x}{the ace.test object to print}
}
\value{
a rounded ace.test object

NULL only outputs to CAT
}
\description{
A S3 function to produce a summary 
  of the results of an ace.test.

A S3 function to produce a summary 
  of the results of an ace.test.
}
