% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ace.test.R
\name{summary.acetest}
\alias{summary.acetest}
\title{ACE permutation test summary}
\usage{
\method{summary}{acetest}(object, ..., digits)
}
\arguments{
\item{object}{the test to summarize}

\item{...}{additional arguments (ignored)}

\item{digits}{Number of significant digits to round too.}
}
\value{
a rounded ace.test object
}
\description{
A S3 function to produce a summary 
  of the results of an ace.test.
}
